python/requirements.txt
mpmath==1.3.0
pycryptodome==3.21.0
sympy==1.13.3
